+++ src/LightMode.ts
@@ -2,6 +2,10 @@
 import DESIGN_TOKENS from "./design-system/tokens.json";
 import { Theme } from "./theme/AppStyles";
 
+// TODO implement some helper(s) to allow declaring key/value pairs as below:
+// backgroundColor: "surface-variant-light,
+
+
 /////////// LIGHT MODE ///////////
 
 const lightColorMode: Theme = {
@@ -18,7 +22,7 @@ const lightColorMode: Theme = {
     // textColor: DESIGN_TOKENS["--md-sys-color-on-secondary-light"],
     backgroundColor: DESIGN_TOKENS["--md-sys-color-surface-variant-light"],
     textColor: DESIGN_TOKENS["--md-sys-color-on-surface-variant-light"],
-    
+
     hoverBackgroundColor: DESIGN_TOKENS['--md-sys-color-primary-light'],
     hoverTextColor: DESIGN_TOKENS["--md-sys-color-on-primary-light"],
     activatedBackgroundColor: DESIGN_TOKENS["--md-sys-color-secondary-light"],
@@ -124,6 +128,8 @@ const lightColorMode: Theme = {
         backgroundColor: DESIGN_TOKENS['--md-sys-color-secondary-container-light'],
         color: DESIGN_TOKENS['--md-sys-color-on-secondary-container-light'],
         // RELEASE ITEM
+        // TODO: go from Light to Darker when opening dialogs.
+        // TODO clearly declare here General, Buttons, Dialogs Background interface
         item: {
           // backgroundColor: DESIGN_TOKENS['--md-sys-color-secondary-container-light'],
           // --md-sys-color-surface-variant-light
@@ -131,6 +137,47 @@ const lightColorMode: Theme = {
           color: DESIGN_TOKENS['--md-sys-color-on-secondary-container-light'],
           onHoverBackgroundColor: DESIGN_TOKENS['--md-sys-color-tertiary-container-light'],
           onHoverColor: DESIGN_TOKENS['--md-sys-color-on-tertiary-container-light'],
+          // <pre> / <code> Colors
+          codeColor: DESIGN_TOKENS['--md-sys-color-on-secondary-container-light'],
+
+          // CODE BACKGROUND, should better be a bit "darker" than the Dialog Background, for better contrast
+
+          // Nice Light Grey Color in general. Darker than the Dialog Background. Good-enough, for now, it would be nicer to redesign this.
+          // this is a nice "mellow" yellow background; definitely darker than the dialog. can work!
+          codeBackgroundColor: DESIGN_TOKENS['--md-sys-color-tertiary-container-light'],
+          // this is a very nice darker "version of pink" compared to the dialog background "pink". very nice!
+          // codeBackgroundColor: DESIGN_TOKENS['--md-sys-color-outline-variant-light'],
+          // Not bad, redish color darker than the Dialog, but not too dark. Passable solution.
+          // codeBackgroundColor: DESIGN_TOKENS['--md-sys-color-inverse-primary-light']
+
+
+          // codeBackgroundColor: DESIGN_TOKENS['--md-sys-color-surface-variant-light'], // SAME AS Dialog Background
+          // codeBackgroundColor: DESIGN_TOKENS['--md-sys-color-on-secondary-light'], // WHITE
+          // codeBackgroundColor: DESIGN_TOKENS['--md-sys-color-inverse-on-surface-light'],
+          // very brownish/greenish and too dark, for this case.
+          // codeBackgroundColor: DESIGN_TOKENS['--md-sys-color-tertiary-light'],
+          // Very Nice Dark color, but too dark for this case. Use elsewhere, it is very nice.
+          //  codeBackgroundColor: DESIGN_TOKENS['--md-sys-color-secondary-light']
+          // same (or almsot the same) as the Dialog Background. no contrast at all
+          // codeBackgroundColor: DESIGN_TOKENS['--md-sys-color-primary-container-light']
+          // Completely white, much lighter than Dialog Background
+          // codeBackgroundColor: DESIGN_TOKENS['--md-sys-color-on-primary-light']
+          // belongs to the Dark family of tones. too dark for this case. but viable for varying dark elsewhere.
+          // codeBackgroundColor: DESIGN_TOKENS['--md-sys-color-inverse-surface-light']
+          // this is white
+          // codeBackgroundColor: DESIGN_TOKENS['--md-sys-color-inverse-on-surface-light']
+          // nice in between grey and black, but too dark for here
+          // codeBackgroundColor: DESIGN_TOKENS['--md-sys-color-outline-light']
+          // belongs to the Dark family of tones. too dark for this case. but viable for varying dark elsewhere.
+          // codeBackgroundColor: DESIGN_TOKENS['--md-sys-color-on-surface-variant-light']
+          // this is white
+          // codeBackgroundColor: DESIGN_TOKENS['--md-sys-color-inverse-on-surface-light']
+          // this is the slightest possible "darker" than the dialog background. it does not offer practical contrast.
+          // codeBackgroundColor: DESIGN_TOKENS['--md-sys-color-secondary-container-light']
+          
+          onHoverCodeBackgroundColor: DESIGN_TOKENS['--md-sys-color-inverse-primary-light'],
+          dialogBackgroundColor: DESIGN_TOKENS['--md-sys-color-surface-variant-light'],
+
         },
       },
     },